refresh logic implemented. todo - currently refresh requires two taps to show changes. need to fi
