+++ app/src/main/java/com/example/support/member/MemberFragment.kt
@@ -4,16 +4,20 @@ package com.example.support.member
 import android.os.Bundle
 import android.view.*
 import androidx.fragment.app.Fragment
+import androidx.fragment.app.activityViewModels
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.fragment.findNavController
 import com.example.support.R
+import com.example.support.RefreshViewModel
 import com.example.support.databinding.MemberFragmentBinding
 import com.example.support.networking.TeamMember
 import com.squareup.picasso.Picasso
 
 class MemberFragment : Fragment() {
     private lateinit var binding: MemberFragmentBinding
+    private val refreshViewModel: RefreshViewModel by activityViewModels()
+    private lateinit var viewModel : MemberViewModel
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
@@ -24,11 +28,10 @@ class MemberFragment : Fragment() {
         binding.lifecycleOwner = this
         val teamMember: TeamMember = MemberFragmentArgs.fromBundle(requireArguments()).selectedProperty
         val viewModelFactory = MemberViewModelFactory(teamMember, application)
-        val viewModel = ViewModelProvider(
+        viewModel = ViewModelProvider(
             this, viewModelFactory
         ).get(MemberViewModel::class.java)
         binding.viewModel = viewModel
-        Picasso.with(application).load(viewModel.selectedProperty.value?.imgSrcUrl).into(this.binding.mainPhotoImage)
         return binding.root
     }
 
@@ -53,7 +56,7 @@ class MemberFragment : Fragment() {
             return false
         }
         else if(item.itemId == R.id.refresh_action){
-            //Todo - handle refresh action logic here
+            viewModel.updateSelectedProperty(refreshViewModel.refreshData())
             return true
          }
