+++ app/src/main/res/layout-land/team_fragment.xml
@@ -7,8 +7,8 @@
     <data>
 
         <variable
-            name="viewModel"
-            type="com.example.support.team.TeamViewModel" />
+            name="refreshViewModel"
+            type="com.example.support.RefreshViewModel" />
 
         <import type="android.view.View" />
 
@@ -28,15 +28,15 @@
             app:layout_constraintRight_toRightOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
-            app:listData="@{viewModel.members}"
+            app:listData="@{refreshViewModel.members}"
             tools:listitem="@layout/team_member_view" />
 
         <ImageView
             android:id="@+id/status_image"
             android:layout_width="0dp"
             android:layout_height="0dp"
-            android:visibility="@{viewModel.status.visible ? View.VISIBLE : View.GONE}"
-            app:imageResource="@{viewModel.status.imageResource}"
+            android:visibility="@{refreshViewModel.status.visible ? View.VISIBLE : View.GONE}"
+            app:imageResource="@{refreshViewModel.status.imageResource}"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
