+++ app/src/main/res/layout/member_fragment.xml
@@ -36,7 +36,7 @@
             android:layout_marginEnd="36dp"
             android:layout_marginBottom="36dp"
             android:contentDescription="TODO"
-            app:imageResource="@{viewModel.displayAvailability.intValue()}"
+            app:imageResource="@{viewModel.displayAvailability}"
             app:layout_constraintBottom_toBottomOf="@+id/main_photo_image"
             app:layout_constraintEnd_toEndOf="@+id/main_photo_image" />
